fix: JWT refresh

+++ server/src/files/files.service.ts
@@ -1,17 +1,44 @@
-import { Injectable } from "@nestjs/common";
+import {
+    Injectable,
+    Logger,
+    NotFoundException,
+    UnauthorizedException,
+} from "@nestjs/common";
 import { DatabaseService } from "src/database/database.service";
 import { FileUploadDto } from "./dto/fileUpload.dto";
+import { join } from "path";
+import { promisify } from "util";
+import { unlink } from "fs";
+import { Response } from "express";
 
 @Injectable()
 export class FilesService {
     constructor(private db: DatabaseService) {}
 
+    private readonly logger = new Logger("FilesService");
+    private readonly unlinkAsync = promisify(unlink);
+
     async findAll(id: number) {
         return this.db.file.findMany({
             where: {
-                userId: id,
+                userId: +id,
             },
         });
+    }
+
+    async getFileByFileName(userId: number, filename: string, res: Response) {
+        const file = await this.db.file.findUnique({
+            where: {
+                userId: +userId,
+                filename: filename,
+            },
+        });
+
+        if (!file) throw new UnauthorizedException();
+
+        const filePath = join(__dirname, "..", "..", "uploads", filename);
+
+        return res.sendFile(filePath);
     }
 
     async uploadFile(fileUploadDto: FileUploadDto, userId) {
@@ -27,19 +54,55 @@ export class FilesService {
     }
 
     async deleteFileById(userId: number, fileId: number) {
+        const file = await this.db.file.findUnique({
+            where: {
+                id: +fileId,
+                userId: +userId,
+            },
+        });
+        if (!file) throw new NotFoundException("File not found");
+
+        await this.deleteFileInStorage(file);
+
         return this.db.file.delete({
             where: {
-                userId: userId,
-                id: fileId,
+                id: +fileId,
+                userId: +userId,
             },
         });
     }
 
     async deleteAllFiles(userId: number) {
+        // const filesCount = await this.db.file.count({
+        //     where: {
+        //         userId: +userId,
+        //     },
+        // });
+        // if (filesCount === 0) throw new NotFoundException("Files not found");
+
+        const files: any = await this.db.file.findMany({
+            where: {
+                userId: +userId,
+            },
+        });
+
+        for (const file of files) {
+            this.logger.error(file);
+            await this.deleteFileInStorage(file);
+        }
+
         return this.db.file.deleteMany({
             where: {
-                userId: userId,
+                userId: +userId,
             },
+        });
+    }
+
+    async deleteFileInStorage(file: any) {
+        const filePath = join(__dirname, "..", "..", "uploads", file.filename);
+
+        await this.unlinkAsync(filePath).catch((err) => {
+            console.error("Error deleting file:", err);
         });
     }
 }
